fix(deposit): remove relations option to fix PostgreSQL FOR UPDATE error
The error 'FOR UPDATE cannot be applied to the nullable side of an outer join'
was caused by combining pessimistic_write lock with relations: ['user'].

PostgreSQL doesn't allow FOR UPDATE with LEFT JOINs.

We don't need to load the user relation anyway since:
1. deposit.userId is already a column on the table
2. We load the user separately later with its own lock

This should finally fix deposits not being auto-confirmed.